+++ src/client/iam_private/CMakeLists.txt
@@ -1,13 +1,18 @@
 _ydb_sdk_add_library(client-iam_private)
 
-target_link_libraries(client-iam_private PUBLIC
-  yutil
-  api-client-yc_private
-  client-iam-common
+target_link_libraries(client-iam_private
+  PUBLIC
+    client-ydb_types-credentials
+    library-jwt
+    yutil
+  PRIVATE
+    api-client-yc_private
+    client-iam-common
 )
 
-target_sources(client-iam_private PRIVATE
-  iam.cpp
+target_sources(client-iam_private
+  PRIVATE
+    iam.cpp
 )
 
 _ydb_sdk_make_client_component(IamPrivate client-iam_private)
