+++ src/client/iam_private/iam.cpp
@@ -1,6 +1,6 @@
 #include <ydb-cpp-sdk/client/iam_private/iam.h>
 
-#include <ydb-cpp-sdk/client/iam/common/iam.h>
+#include <src/client/iam/common/iam.h>
 
 #include <src/api/client/yc_private/iam/iam_token_service.pb.h>
 #include <src/api/client/yc_private/iam/iam_token_service.grpc.pb.h>
